👨‍🎨 Adjust layout for small sized screens
Adjust layout so that the `w-full` in combination with the `mx` don't
cause the content to be wider than the screen on small devices.

There's maybe a better way to do this with padding but with this set up
the layout makes sure that short content will take up the full prose
width on larger screens and long content does not overflow off the
screen on smaller devices!

Yay!