Add @tailwind/typography to style articles
Add the `@tailwind/typography` plugin to handle styling my articles.
This was much much easier than coming up with my own typography rules
and handles a lot of situations I would not have thought of.

Adds a small fix with the `prose-pre:bg-[#f0f0f0]` so that the
background of the `pre` tags matches the `code` blocks that are styled
by make up.