Add Blog.next_article/1 & Blog.previous_article/1
Add the `Blog.next_article/1` and `Blog.previous_article/1` functions to
enable easily accessing the next and previous articles for the provided
article.

This function will allow me to add some nice `Next` and `Previous`
navigation buttons to the Article show UI.

I have a couple notes on this work.

---
Testing

Right now the tests are using the production articles, which is fine I
guess.

I really wanted them to use some test fixtures but there was something
going on with builds and having Elixir include test articles at
`test/support/fixtures/` in the test build didn't work as I expected it
to.

I had expected that since `mix.exs` has the following config:
```
defp elixirc_paths(:test), do: ["lib", "test/support"]
```
That the files placed into the `support` directory would be compiled
into the build. I added an `@external_resource` to the `Blog` context to
mark the `.md` files as an external resource and this didn't help.

After spending some time on this I moved on from it and will likely ask
the forums.

---
Efficiency

The `next_article/1` and `previous_article/1` could probably be more
efficient but that would come at the cost of clarity.

Right now they'll do 2 passes through the list of articles.

If this was a large list I would be worried but I honestly don't see
myself writing so many articles this becomes an issue...