Reconfigure Ecto
Reconfigure Ecto so that there is an associated database for this
project.

With a database in place I can begin to add functionality that will
allow for some kind of analytics on my website.
I'd like to do this without paying and with the minimum amount of
tracking of page visitors as possible.

Inspired by [this Dashbit
article](https://dashbit.co/blog/homemade-analytics-with-ecto-and-elixir)
I decided to roll my own analytics based on their approach.