Add Metric struct
Add the Metric struct along with its migration and context.

A composite primary key is used for the Metric as insertions/updates
will be done based upon the date and path. Which means an easy method of
accessing the Metric based on the date and path is needed.

It is also expected that there is a unique row in the database for each
date and path combination.

Inspired by [this Dashbit
article](https://dashbit.co/blog/homemade-analytics-with-ecto-and-elixir)
I decided to roll my own analytics based on their approach.