Add the Metrics Supervisor and Worker
Add the Metrics Supervisor and Worker to handle the persistance of
Metrics in a separate process.

This allows the tracking of Metrics to be done separately from the
rendering of the page's content and means that any failures to store the
metrics in the database will not impact the user's ability to view the
content.

Inspired by [this Dashbit
article](https://dashbit.co/blog/homemade-analytics-with-ecto-and-elixir)
I decided to roll my own analytics based on their approach.