Add track_metrics plug
Add the `track_metrics/2` plug to handle tracking page views.

This simple plug will kick off the tracking of metrics in a separate
process if the conn response is a 200 status code.

With this in place, a basic metrics system exists for this website that
does not require cookies. Keeping the actual tracking of the users low.

Inspired by [this Dashbit
article](https://dashbit.co/blog/homemade-analytics-with-ecto-and-elixir)
I decided to roll my own analytics based on their approach.