Add basic metrics dashboard
Add a (very) basic metrics dashboard that will allow viewing the metrics
for my website!

To begin with this is a simple list of the Dates and paths. The screen
will render _all_ the metrics in the database.

This is a first pass at this dashboard and as such there's some code
that _might_ need cleaning up such as making the test's metric fixture
reusable, removing the call to the Repo module in the controller, and
other small things like that.

In the spirit of incremental and quick releases, I decided to just push
this live as it is since it works and allows me to see my site's
metrics!