Update Metrics index with weekly metrics
Update the Metrics index to display the weekly metrics.

This is a first pass on showing "useful" metrics, some thought needs to
be put into what would be nice to see on this screen.

Likely, it should be something like most popular pages for week, month,
all time or something like that.

For now it's just what's gone on this week.