Add article about JS packages in Phoenix LiveViews
Add article about using JS packages in Phoenix LiveViews.

This was prompted by a small experiment I did that required adding a
Leaflet map to a LiveView.

I found the documentation on how to do this to be somewhat misleading
and it took me much longer to get Leaflet installed along with its
styles correctly.

Hopefully this article will help others.