Adjust Fly.io deployment configuration
Adjust the Fly.io deployment configuration so that it is possible to
connect to te production application with Livebook.

This is the end of a long effort in trying to connect a local Livebook
desktop app to a Production app on Fly.io.

It's an amalgamation of [this connect to Livebook
doc](https://fly.io/docs/elixir/advanced-guides/connect-livebook-to-your-app/),
[this do interesting things with Livebook
doc](https://fly.io/docs/elixir/advanced-guides/interesting-things-with-livebook/),
and [this thread asking for
help](https://community.fly.io/t/unable-to-connect-local-livebook-desktop-to-production-application/17719/2).

In the end, issues with connecting were due to a missing
`rel/env.sh.eex` file that configures some important environment
variables in the Fly deployments.
The _really_ important one is the `RELEASE_NODE` environment variable as
that names the node and is what can be used, along with the
`RELEASE_COOKIE` to connect to the node.