Add Metrics Livebook app
Add a very simple Metrics Livebook app and its deployment process.

This is step one to building out a very simple metrics dashboard for my
website.

Since Livebook has support for charts etc. out of the box, this will
allow me to build out a simple page that display the metrics I am
tracking in a nice way quite quickly, which is great!