Start Livebook analytics app article
Start an article about using Livebook as an analytics visualizer.

This is a rough draft with portions filled out that go over how to
create visualizations in Livebook and then deploy it to Fly.io.

Hopefully it will help other people looking for Livebook ideas!

As part of this artivle the metrics livebook was changed to be more
functional and have an actual chart that allows selecting a date range
and viewing the results.