Add to Livebook Analytics blog post
Add the the Livebook Analytics blog post. This is getting really long
winded and I'm not sure what the point of it is anymore.

I decided to split the last section of this article into its own shorter
post about deploying a Livebook App to Fly.io and come back around to
this and see if there's something a little more focussed to write.

Currently it's reading like a very long "here's a thing I did" and
there's not much compelling about it IMO.