Convert main website pages to LiveView
Convert the main website pages to LiveView as I would like to begin
adding some slick new behaviour to these screens.

This migration required some adjustments to the app layout to made in
order to support the dynamic page views counter.

With the swap to LiveViews, the main pages are ran inside of a
`live_session` block.
This means that when navigating between the `/`, `/now`, and `/about`
pages the Plug pipeline is called exactly once.
Normally this would not be an issue but since the `page_visits`
component gets its assigns from the Plug pipeline, it meant that the
page visits count would never be updated when navigating between the
pages and the wrong count would be displayed.

In order to resolve this two things needed to be done.

First a LiveView hook was added that sets the `page_visits` assign in
the `handle_params/3` callback.
The `handle_params/3` callback is used as it is the only time in the
lifecycle that the URI is available and the Metrics are counted based on
the URI path.

Second, the `app.html.heex` template was updated to house the `header`,
`main`, and `footer` elements from the `root.html.heex`.
This was done as the `root.html.heex` is outside of the DOM that
LiveView has access to.
Thus, if we want to be able to dynamically change the
`Layout.page_visits/1` display state, it needs to be housed inside of
the DOM that LiveView has access to.
I don't _love_ this but it works so 🤷‍♂️