Fix LiveView Metric tracking
Fix LiveView Metric tracking so that LiveViews in the same
`live_session/1` actually track their page views.

In [this previous
commit](https://github.com/SamuelWillis/samuelwillis.dev/commit/a472527e336a71402d060ea5a1166c4bb7ab4a1e)
a fix was applied to rendering the page view counter, however, it was
missed that the same issue applies with tracking the Metrics as well.

Namely, when LiveViews are in the same `live_session/1` the plug
pipeline happens exactly once for the initial render of the initial
LiveView.
After that the Plug pipeline is not hit which means that every page load
after the initial LiveView does not have its metrics tracked.

I decided to solve this in a similar fashion as the page counter via the
introduction of a Hook.
This hook uses the `start_async/4` function to kick off the metric
tracking in the background, similar to the Plug.
The `start_async/4` only runs on the connected

There's a couple downsides to this approach.

First, the Plug + Hook cannot be assigned together as they will double
count metrics.
This requires manual addition of the correct one to each place.
It also means that a generic Plug version cannot be slapped into the
pipeline and just work.

Secondly, the metrics are now only counted on the connected render.
Which could miss some Metrics.