Add loading spinner to HomeLive
Add a loading spinner to HomeLive so that it does not attempt to load
two images and take a long time to connect the LiveView.

This will prevent the static render from loading an image and blocking
the LiveView connection until said image is loaded.

Once the LiveView is connected an img will be rendered on the screen.

Locally this was prone to a flash of unstyled content which I really
don't like but I am curious if that's not a thing when deployed.