Get websockets working on Fly + custom domain
Fix websocket connections for custom domains.

Without this adjustments websockets will not connect successfully on
domains that do not match the domain specified in the `fly.toml`.

This meant that the custom domain configured was unable to connect to
the websocket and, thus, LiveView was not working as expected.

Setting `check_origin: :conn` allows any origin matching the request
connection's host, port, and scheme.
[Docs](https://hexdocs.pm/phoenix/Phoenix.Endpoint.html#socket/3-common-configuration).

Configuring the `force_ssl` allows us to specify what headers should be
used when rewriting a HTTP request to HTTPS.
This is handy as our application is behind Fly's load balancer and SSL
termination.
* `hsts: true`: declares the site is accessible via HTTPS only.
* `host: nil`: allows dynamic redirects to the host of the request
  enforcing HTTPS on the fly domain and any custom domains.
* `rewrite_on: [...]`: specifies the headers to use when rewriting http
  to https

Based on [this community
thread](https://community.fly.io/t/phoenix-liveview-constantly-refreshes-with-custom-domain/3384/9)